backend/requirements.txt
Django==3.0
psycopg2-binary==2.8.5
djangorestframework==3.11.0
django-rest-swagger==2.2.0
pdbpp==0.10.2
factory-boy==2.12.0
